\name{Rbowtie-package}
\alias{Rbowtie-package}
\alias{Rbowtie}
\docType{package}

\title{R bowtie wrapper}

\description{
This package provides an R wrapper around the popular bowtie short read aligner and around SpliceMap, a de novo splice junction discovery and alignment tool.
}

\details{
See packageDescription('Rbowtie') for package details.
}

\author{
Florian Hahne, Anita Lerch, Michael B Stadler
}

%\references{TODO}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
  \code{\link{bowtie}}, \code{\link{SpliceMap}}
}

\examples{
\dontrun{
    example(bowtie)
    example(SpliceMap)
}
}
